+++ examples/resources/nxos_feature_bfd/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_bfd.example "sys/fm/bfd"