+++ internal/provider/data_source_nxos_ospf_interface_test.go
@@ -34,28 +34,36 @@ func TestAccDataSourceNxosOSPFInterface(t *testing.T) {
 
 const testAccDataSourceNxosOSPFInterfacePrerequisitesConfig = `
 resource "nxos_rest" "PreReq0" {
+  dn = "sys/fm/bfd"
+  class_name = "fmBfd"
+  content = {
+      adminSt = "enabled"
+  }
+}
+
+resource "nxos_rest" "PreReq1" {
   dn = "sys/ospf"
   class_name = "ospfEntity"
   content = {
   }
 }
 
-resource "nxos_rest" "PreReq1" {
+resource "nxos_rest" "PreReq2" {
   dn = "sys/ospf/inst-[OSPF1]"
   class_name = "ospfInst"
   content = {
       name = "OSPF1"
   }
-  depends_on = [nxos_rest.PreReq0, ]
+  depends_on = [nxos_rest.PreReq1, ]
 }
 
-resource "nxos_rest" "PreReq2" {
+resource "nxos_rest" "PreReq3" {
   dn = "sys/ospf/inst-[OSPF1]/dom-[default]"
   class_name = "ospfDom"
   content = {
       name = "default"
   }
-  depends_on = [nxos_rest.PreReq1, ]
+  depends_on = [nxos_rest.PreReq2, ]
 }
 
 `
@@ -75,7 +83,7 @@ resource "nxos_ospf_interface" "test" {
   network_type = "p2p"
   passive = "enabled"
   priority = 10
-  depends_on = [nxos_rest.PreReq0, nxos_rest.PreReq1, nxos_rest.PreReq2, ]
+  depends_on = [nxos_rest.PreReq0, nxos_rest.PreReq1, nxos_rest.PreReq2, nxos_rest.PreReq3, ]
 }
 
 data "nxos_ospf_interface" "test" {